Add SC16 tutorial to the Spack docs.
