+++ lib/spack/docs/tutorial/.gitattributes
@@ -1,0 +1,1 @@
+Spack-SC16-Tutorial.pdf filter=lfs diff=lfs merge=lfs -text
